Remove SettingDefinitions
